requirements.txt
# SkillMCP-Agent Dependencies

# ===== 核心框架 =====
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ===== HTTP 客户端 =====
httpx>=0.26.0
aiohttp>=3.8.0

# ===== 向量存储 =====
faiss-cpu>=1.7.4
numpy>=1.24.0

# ===== 文档处理 =====
pypdf>=3.0.0
python-docx>=0.8.11

# ===== 数据处理 =====
tiktoken>=0.5.0
python-multipart>=0.0.9

# ===== 日志 =====
loguru>=0.7.0

# ===== 测试 =====
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# ===== 开发工具 =====
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0

# ===== 可选：OpenAI =====
# openai>=1.12.0

# ===== 可选：本地 Embedding =====
# sentence-transformers>=2.3.0
# torch>=2.0.0
